Added support for "versioning"
The versioning system allows your application to specify a desired version with as much precision as it needs. By dropping version fields, you end up wild carding a field. For instance, consider a set of versions:

 1.9.1
 1.8.4
 1.8.2

A version of "1.8.2" select the obvious, fully-specified version. Specifying a version of "1.8" would select 1.8.4  since this is the highest version released in the 1.8 branch. For much the same reason, a request for "1" loads version 1.9.1, since this is the highest version released in the 1 branch. Note, these versioning semantics work the same way when using google.load and when using direct script urls.